add side nav data & fix repository issues and update some things on base repo & add all cart functionalities

+++ OnlineShop.Core/Entities/Product.cs
@@ -10,7 +10,7 @@ namespace OnlineShop.Core.Entities
 
         public string? Description { get; set; }
         public decimal Price { get; set; }
-        public float discount { get; set; }
+        public decimal discount { get; set; }
 
         public string CategoryId { get; set; }
         public Category? Category { get; set; }
